Create RotM001Script
I wrote this script because I couldn't seem to wrap my head around the Story manager in the Creation Kit. Originally intended to launch the next quest in the chain, as the mod grew so did the script. Additionally since I can't leave well enough alone I decided it was necessary to have this script clean up the persistent references to prevent save game bloat.