Create RotM007SimMasterScript
The VR simulations are controlled by a small network of terminals, five in total. The Master Terminal, which allows the Player to select a Map/Quest, with the four sub terminals allowing the Player to select the desired enemy type. This is the Master terminal script for Map/Quest Selection.